Add files via upload
We start by initializing the dimensions of the input matrix mat as m (number of rows) and n (number of columns). This will be used later in the code.

We create a result matrix called result of size m x n to store the distance of the nearest 0 for each cell. Initially, we set all the values in result to a large value (in this case, INT_MAX) to represent an unreachable distance.

We create a queue called q to store the cell indices that need to be processed during the breadth-first search.

We iterate through the input matrix mat and enqueue the indices of all the cells with value 0 into the queue. Additionally, we mark the distance of these cells in the result matrix as 0. This initialization step ensures that the cells with value 0 are considered as starting points for the breadth-first search.

Now, we perform the breadth-first search using the queue q. While the queue is not empty, we process the cells in a breadth-first manner.

For each cell (row, col) popped from the queue, we check its neighboring cells in four directions: up, down, left, and right. We define the directions as a vector of pairs called directions, where each pair represents the change in row and column indices to move in a specific direction.

If the neighboring cell indices (newRow, newCol) are within the valid range of the matrix and the current distance in the result matrix for (newRow, newCol) is greater than the distance of the current cell plus 1, we update the distance in the result matrix and enqueue the neighboring cell (newRow, newCol) into the queue for further processing.

Once the breadth-first search is completed, the result matrix will contain the distance of the nearest 0 for each cell in the input matrix.

Finally, we return the result matrix.